# backend/target/generated-sources/openapi/src/main/java/com/hotelapp/backend/api/ReservationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.hotelapp.backend.api;

import com.hotelapp.backend.api.model.ReservationDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-17T22:59:14.981834800+01:00[Europe/Berlin]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Reservation", description = "the Reservation API")
public interface ReservationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /reservation : Create a new reservation
     *
     * @param reservationDto Reservation to create (required)
     * @return Reservation created successfully (status code 201)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createReservation",
        summary = "Create a new reservation",
        tags = { "Reservation" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Reservation created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ReservationDto.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/reservation",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ReservationDto> createReservation(
        @Parameter(name = "ReservationDto", description = "Reservation to create", required = true) @Valid @RequestBody ReservationDto reservationDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"roomNumber\" : 6, \"checkOutDate\" : \"2000-01-23\", \"id\" : 0, \"guestEmail\" : \"guestEmail\", \"checkInDate\" : \"2000-01-23\", \"guestName\" : \"guestName\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /reservation/{id} : Delete a reservation
     *
     * @param id  (required)
     * @return Reservation deleted (status code 204)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteReservation",
        summary = "Delete a reservation",
        tags = { "Reservation" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Reservation deleted"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/reservation/{id}"
    )
    
    default ResponseEntity<Void> deleteReservation(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /reservations : Get all reservations
     *
     * @return A list of reservations (status code 200)
     */
    @Operation(
        operationId = "getAllReservations",
        summary = "Get all reservations",
        tags = { "Reservation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of reservations", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ReservationDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/reservations",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ReservationDto>> getAllReservations(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"roomNumber\" : 6, \"checkOutDate\" : \"2000-01-23\", \"id\" : 0, \"guestEmail\" : \"guestEmail\", \"checkInDate\" : \"2000-01-23\", \"guestName\" : \"guestName\" }, { \"roomNumber\" : 6, \"checkOutDate\" : \"2000-01-23\", \"id\" : 0, \"guestEmail\" : \"guestEmail\", \"checkInDate\" : \"2000-01-23\", \"guestName\" : \"guestName\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /reservation/{id} : Get a reservation by ID
     *
     * @param id  (required)
     * @return Reservation found (status code 200)
     *         or Reservation not found (status code 404)
     */
    @Operation(
        operationId = "getReservationById",
        summary = "Get a reservation by ID",
        tags = { "Reservation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Reservation found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ReservationDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Reservation not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/reservation/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ReservationDto> getReservationById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"roomNumber\" : 6, \"checkOutDate\" : \"2000-01-23\", \"id\" : 0, \"guestEmail\" : \"guestEmail\", \"checkInDate\" : \"2000-01-23\", \"guestName\" : \"guestName\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /reservation/{id} : Update a reservation
     *
     * @param id  (required)
     * @param reservationDto Updated reservation data (required)
     * @return Reservation updated (status code 200)
     *         or Reservation not found (status code 404)
     */
    @Operation(
        operationId = "updateReservation",
        summary = "Update a reservation",
        tags = { "Reservation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Reservation updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ReservationDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Reservation not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/reservation/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ReservationDto> updateReservation(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ReservationDto", description = "Updated reservation data", required = true) @Valid @RequestBody ReservationDto reservationDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"roomNumber\" : 6, \"checkOutDate\" : \"2000-01-23\", \"id\" : 0, \"guestEmail\" : \"guestEmail\", \"checkInDate\" : \"2000-01-23\", \"guestName\" : \"guestName\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}